Update and get useBackButtonListener to a somewhat useable level
