+++ src/hooks/useBackButtonListener/useBackButtonListener.js
@@ -1,19 +1,20 @@
 import { useEffect } from 'react';
-import useDispatch from './useDispatch';
+import useHistoryBlock from './useHistoryBlock';
 
-//@TODO: Probs clean this up later
 /**
  * Prevent the user from hitting the back button.
  */
 const useBackButtonListener = () => {
-    const dispatch = useDispatch();
+  const location = sessionStorage.getItem('LOCATION');
 
-    return useEffect(() => {
-        window.addEventListener('popstate', dispatch);
-        return () => {
-            window.removeEventListener('popstate', dispatch);
-        }
-    }, [dispatch]);
+  const dispatch = useHistoryBlock(location);
+
+  return useEffect(() => {
+    window.addEventListener('popstate', dispatch);
+    return () => {
+      window.removeEventListener('popstate', dispatch);
+    };
+  }, [dispatch]);
 };
 
 export default useBackButtonListener;
